requirements.txt
# Additional requirements go here
Django>=2.2,<=3.1
djangorestframework
pytils
